requirements-api.txt
Flask==3.1.0
joblib==1.4.2
numpy==2.2.1
Pillow==11.0.0
scikit-learn==1.6.1
